Improve convenience
Handle AppBar sizing when shell
Use ABEdge enum in TrayHostSizeData
Add MouseHelper for tray icon cursor position

+++ src/ManagedShell.WindowsTray/NotificationArea.cs
@@ -21,7 +21,13 @@ namespace ManagedShell.WindowsTray
         };
 
         const string VOLUME_GUID = "7820ae73-23e3-4229-82c1-e41cb67d5b9c";
-        readonly NativeMethods.Rect defaultPlacement = new NativeMethods.Rect { Top = 0, Left = GetSystemMetrics(0) - 200, Bottom = 23, Right = 23 };
+        readonly NativeMethods.Rect defaultPlacement = new NativeMethods.Rect
+        {
+            Top = 0,
+            Left = GetSystemMetrics(0) - 200,
+            Bottom = 23,
+            Right = 23
+        };
         
         public string[] PinnedNotifyIcons { get; internal set; }
         public IntPtr Handle { get; private set; }
@@ -32,7 +38,17 @@ namespace ManagedShell.WindowsTray
         private TrayHostSizeDelegate trayHostSizeDelegate;
         private object _lockObject = new object();
         private ShellServiceObject shellServiceObject;
-        private TrayHostSizeData trayHostSizeData = new TrayHostSizeData { edge = (int)ABEdge.ABE_TOP, rc = new NativeMethods.Rect { Top = 0, Left = 0, Bottom = 23, Right = GetSystemMetrics(0) } };
+        private TrayHostSizeData trayHostSizeData = new TrayHostSizeData
+        {
+            edge = ABEdge.ABE_TOP,
+            rc = new NativeMethods.Rect
+            {
+                Top = 0,
+                Left = 0,
+                Bottom = 23,
+                Right = GetSystemMetrics(0)
+            }
+        };
 
         public NotificationArea(TrayService trayService, ExplorerTrayService explorerTrayService) : this(DEFAULT_PINNED, trayService, explorerTrayService)
         {
